fix test against NSNotFound
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/ec/trunk@35648 72102866-910b-0410-8b05-ffd578937521

+++ EcCommand.m
@@ -2248,10 +2248,10 @@ static NSString*	cmdWord(NSArray* a, unsigned int pos)
 
   if (o != nil)
     {
-      NSString	*m;
-      unsigned	i;
-      BOOL	transient = [o transient];
-      NSString	*name = [[[o name] retain] autorelease];
+      NSString	        *m;
+      NSUInteger	i;
+      BOOL	        transient = [o transient];
+      NSString	        *name = [[[o name] retain] autorelease];
 
       m = [NSString stringWithFormat: 
 	@"\n%@ removed (unregistered) server -\n  '%@' on %@\n",
@@ -2282,10 +2282,10 @@ static NSString*	cmdWord(NSArray* a, unsigned int pos)
 
   if (o)
     {
-      NSString	*m;
-      unsigned	i;
-      BOOL	transient = [o transient];
-      NSString	*name = [[[o name] retain] autorelease];
+      NSString	        *m;
+      NSUInteger       	i;
+      BOOL	        transient = [o transient];
+      NSString	        *name = [[[o name] retain] autorelease];
 
       m = [NSString stringWithFormat: 
 	@"\n%@ removed (unregistered) server -\n  '%@' on %@\n",
